Report results properly to test.py from driver.js.
If the PDFDoc creation failed in driver.js then that was not informed back to
test.py. This lead to State.remaining being off by one. And that did not let
the test to end as expected. Instead the test hung indefinitely.

This change now reveals TEST-UNEXPECTED-FAIL which was hidden previously.