Re-enable ‘open file’ button in Chrome.
Originally disabled as a way to not violate the ‘inline script/style’ not allowed permission. We resolve the inline styling issues by moving styling to CSS. Also only accept PDF documents in the file dialog (yes it’s a weak form of filtering since not every OS/browser enforces it, but when it’s enabled it’ll improve the UX).