Remove custom |pagerender| event from page_view.js, except for |GENERIC| builds
The custom `pagerender` event in page_view.js is not used anywhere in the codebase.
PR 5295 introduced the (in my opinion) better named `pagerendered` event, which is placed in pdf_viewer.js and is actually being used.
Hence this patch removes `pagerender`, since I believe it to be unnecessary code duplication.

Note: for backward compatibility, the event is still present in `GENERIC` builds.