Don't map glyphs to Unicode combined Thai character (bug1046314)
This is a very tentative patch, that fixes https://bugzilla.mozilla.org/show_bug.cgi?id=1046314.
Even though the Thai Unicode character `0x0E33` has a length of `1`, when copied I noticed that it's made up of "two" characters. I'm assuming that this is what's causing the problem in PDF file in the bug, hence I'm thinking that we can simply ignore that spot when creating the charCode map.