Rip out the current implementation of `PDFHistory`
The current implementation of `PDFHistory` contains a number of smaller bugs, which are very difficult to address without breaking other parts of its code.
Probably the main issue with the current implementation, is that I wrote it quite some time ago, and at the time my understanding of the various edge-cases the code has to deal with was very limited.
Currently `PDFHistory` can still, despite most of those cases being fixed, in some cases lock-up the browser history, essentially preventing the user from navigating back/forward.

Hence rather than trying to iterate on `PDFHistory` to make it better, the easiest approach was to simply replace it entirely with newly written code.