Simplify `PDFLinkService_getDestinationHash` to return a stringified version of the destination array for explicit destinations
Currently for explicit destinations, compared to named destinations, we manually try to build a hash that often times is a quite poor representation of the *actual* destination. (Currently this only, kind of, works for `\XYZ` destinations.)
For PDF files using explicit destinations, this can make it difficult to accurately link to a specific section of the document through the URL.

In practice most PDF files, especially newer ones, use named destinations and these are thus unnaffected by this patch.

With the added, and already existing, type checks in place for destinations, I really don't think that this patch exposes any "sensitive" internal destination code not already accessible through normal hash parameters.

*Please note:* Just trying to improve the algorithm that generates the hash is unfortunately not possible in general, since there are cases where it will simply never work well.
One issue is that `getDestinationHash` currently relies on `_pagesRefCache`, hence it's possible that the hash returned is empty during (ranged/streamed) loading of a PDF file. (In order for the new `PDFHistory` to work, `getDestinationHash` must always return a usable hash.)