Re-write `PDFHistory` from scratch
This patch completely re-implements `PDFHistory` to get rid of various bugs currently present, and to make maintenance easier. Most of the interface is similar to the existing one, but it should be slightly simplified.

The new implementation should be more robust against failure, compared to old one. In the old one, it was too easy to end up in a state which basically caused the browser history to lock-up, preventing the user from navigating back/forward.

Given that the code has to deal with various edge-cases, it's still not as simple as I would have liked, but it should now be somewhat easier to deal with.
The main source of complication for this code is actually that we allow the user to change the hash of a already loaded document. Compared to the old implementation however, this part has been made simpler.

In the new code, I've also added a *lot* more comments (perhaps too many?) to help explain the logic. This is something that the old implementation was serverly lacking, which is a large part of the reason why it was so difficult to maintain.
*Note:* The majority of this patch was actually written a couple of months ago, and then left mostly untouched until a few weeks ago. This was done to ensure that the code, and comments, still made sense to me even when it wasn't fresh in my mind.