Further improvements of the getDownloadInfo and getStats unit tests
It seems nice to ensure that `getStats` not only returns something, but actually returns the expected data when the PDF file has been parsed.

Since I'm touching the file anyway, it seemed like a good idea to make a slight adjustment of the `getDownloadInto` test as well. (I really don't know why I wrote it that way to begin with.)