Stop building the AMO extension
A long time ago it has already been removed from the bot's preview messages, but the extension was still generated (and left unused) in the background. This patch stops building the AMO extension altogether as we do not use and provide it anymore as an add-on. Users that do want to use the extension for highly exceptional cases (such as Android where we have no proper viewer yet) could use the development extension.