Add PDFDataWriter class to generate PDF data
The PDFDataWriter class is a low-level class that supports the raw types
as defined by section 7.3 and 7.5 of the PDF 1.7 spec. It is designed to
support incremental updates in PDF.js.

The immediate use case for developing this class is to add a printing
instruction to the PDF for Chrome, but the functionality can also be
re-used to implement things like persistent annotations.

The code is not referenced in the worker loader yet, because at this
point there is no code that uses the writer. The class has fully been
tested, including a full example of using the class to create a PDF from
scratch (see 'should be able to generate a valid PDF file' in
test/unit/pdf_data_writer_spec.js and test/pdfs/written-with-pdfjs.pdf).