Chromium: Use PDFium for printing if available.
Currently, printing in PDF.js for Chromium is implemented as follows:

1. Render all PDF pages on a <canvas>
2. Send a print request to the browser.

When Chromium receives the print request, it will convert the page to
PDF using Skia, and then open a print preview with PDFium (Chrome's
default PDF reader). This method is suboptimal, since it is essentially
PDF -> HTML and HTML -> PDF.

To enhance the performance, PDF.js will not render the PDF, but send the
PDF to the built-in PDF handler for printing when availability of PDFium
is detected. This brings the following advantages:

- The print job will be started much faster, because PDF.js does not
  have to render the PDF file before triggering the print request.

- Large PDF files can now be printed without out-of-memory errors.

- The generated prints will generally have a better quality (no page
  cut-offs, etc.), because the browser has more control over printer
  settings than PDF.js.

This feature is enabled by default for the GENERIC build, and can
completely be disabled by setting WITH_PRINTING to false. This is also
the first commit that uses editing/pdf_data_writer.js, which was
introduced in an earlier commit.