Add a way to optionally suppress displaying the fallback bar, and logging debug information, when calling `error`
There are some places in the code, e.g. when executing code within a try-catch block, where it would be great to be able to *only* throw an `Error` without bothering the user with the fallback bar.

Considering that `error` is only called when something goes wrong, and that I don't expect that we'll add too many of these special cases, I don't believe that this will slow down the execution time of the code in any noticeable way.