Remove the `hasHtml` data property from annotations
Each annotation type in the PDF specification requires some kind of
rendering and thus an HTML element, so we do not need to specify this
for each annotation class.

This was only used to hide text widget annotations if there is no
content. This is now done with the `showable` property that only applies
to that particular annotation.

Furthermore, we can now remove the if-statement in the `render()`
method. `hasHtml` is gone and `data` is always defined as each
annotation has `data` because each annotation inherits from a base
annotation class that specifies these (see
https://github.com/mozilla/pdf.js/blob/master/src/core/annotation.js#L162).
The check is therefore obsolete.