pagechange Event out-of-range previousPageNumber
If the user navigates to the first/last scene and then tries to navig…
…ate to previous/next scene respectively using the keyboard left/right
arrows, the "pagechange" event is fired but the "previousPageNumber"
parameter is out or the acceptable range (either 0 or "pagesCount + 1").

This fix just ensures that if the navigation is beyond either range, we
just clip it.
This leads to the "pagechange" event still firing but showing that the
"previousPageNumber" and "pageNumber" are the same... logic within the
event-handler should be checking for that anyway.