Add a simple cache in `Catalog_getDestination` to avoid looking up the same destination multiple times
When getting all the destinations (`Catalog.destinations`) the result is cached, to improve performance. However, when individual destinations are fetched, the lookup is repeated every time. Especially for larger PDF files (e.g. the PDF spec), where the `NameTree` containing the destinations can be large, it seems quite unnecessary to repeat the lookup multiple times.